refactor: simplify post-management-page structure
- remove unused hooks and files
- update view component structure
