--- src/views/post-management-page/hooks/index.ts
@@ -1,2 +1,0 @@
-export { useCommentForm, type CommentFormValues } from "./use-comment-form"
-export { usePostForm, type PostFormValues } from "./use-post-form"
