--- src/views/post-management-page/static-model.ts
@@ -1,1 +1,0 @@
-export const LIMIT_OPTIONS = [10, 20, 30] as const
